Crate cbindgen

Structs

Bindings
A bindings header that can be written.

Builder
A builder for generating a bindings header.

Config
A collection of settings to customize the generated bindings.

ConstantConfig
Settings to apply to generated constants.

CythonConfig
Settings specific to Cython bindings.

EnumConfig
Settings to apply to generated enums.

ExportConfig
Settings to apply when exporting items.

FunctionConfig
Settings to apply to generated functions.

LayoutConfig
Settings to apply to generated types with layout modifiers.

MacroExpansionConfig
Settings for custom macro expansion.

MangleConfig
Mangling-specific configuration.

ParseConfig
Settings to apply when parsing.

ParseExpandConfig
Settings to apply when running rustc --pretty=expanded

PtrConfig
Settings to apply to pointers

StructConfig
Settings to apply to generated structs.

Enums

Braces
A style of braces to use for generating code.

DocumentationStyle
How the comments containing documentation should be styled.

Error

ItemType
Different item types that we can generate and filter.

Language
A language type to generate bindings for.

Layout
A type of layout to use when generating long lines of code.

LineEndingStyle
Controls what type of line endings are used in the generated code.

Profile
Controls which Cargo profile is used for macro expansion.

RenameRule
A rule to apply to an identifier when generating bindings.

SortKey
Type which specifies the sort order of functions

Style
A style of Style to use when generating structs and enums.

Constants

VERSION

Functions

generate
A utility function for build scripts to generate bindings for a crate, using a cbindgen.toml if it exists.

generate_with_config
A utility function for build scripts to generate bindings for a crate with a custom config.