Enum cbindgen::RenameRule

pub enum RenameRule {
    None,
    GeckoCase,
    LowerCase,
    UpperCase,
    PascalCase,
    CamelCase,
    SnakeCase,
    ScreamingSnakeCase,
    QualifiedScreamingSnakeCase,
}

A rule to apply to an identifier when generating bindings.

Variants
None

Do not apply any renaming. The default.

GeckoCase

Converts the identifier to PascalCase and adds a context dependent prefix

LowerCase

Converts the identifier to lower case.

UpperCase

Converts the identifier to upper case.

PascalCase

Converts the identifier to PascalCase.

CamelCase

Converts the identifier to camelCase.

SnakeCase

Converts the identifier to snake_case.

ScreamingSnakeCase

Converts the identifier to SCREAMING_SNAKE_CASE.

QualifiedScreamingSnakeCase

Converts the identifier to SCREAMING_SNAKE_CASE and prefixes enum variants with the enum name.

Methods

impl RenameRule

pub fn apply_to_pascal_case(
&self,
text: &str,
context: IdentifierType
) -> String

Applies the rename rule to a string that is formatted in PascalCase.

pub fn apply_to_snake_case(&self, text: &str, context: IdentifierType) -> String

Applies the rename rule to a string that is formatted in snake_case.