Crate cbindgen

Structs

Bindings
A bindings header that can be written.

Builder
A builder for generating a bindings header.

Config
A collection of settings to customize the generated bindings.

ConstantConfig
Settings to apply to generated constants.

EnumConfig
Settings to apply to generated enums.

FunctionConfig
Settings to apply to generated functions.

ParseConfig
Settings to apply when parsing.

StructConfig
Settings to apply to generated structs.

Enums

Braces
A style of braces to use for generating code.

Language
A language type to generate bindings for.

Layout
A type of layout to use when generating long lines of code.

RenameRule
A rule to apply to an identifier when generating bindings.

Constants

VERSION

Functions

generate
A utility function for build scripts to generate bindings for a crate, using a cbindgen.toml if it exists.

generate_with_config
A utility function for build scripts to generate bindings for a crate with a custom config.